requirements.txt
# requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
matplotlib>=3.5.0
numpy>=1.21.0
pymupdf 
pillow 
openpyxl